week8/requirements.txt
Flask
flasgger
pymongo
python-dotenv
PyYAML
PyJWT
Werkzeug
